requirements.txt
flask
python-dotenv
google-cloud-speech
aiohttp
aiohttp-wsgi
pydub